Changed to GWT BigDecimal implementation by KigDecimal

+++ src/main/kotlin/org/ergoplatform/mosaik/LabelFormatter.kt
@@ -18,7 +18,7 @@ object LabelFormatter {
 
             is ErgAmountLabel -> {
                 val ergAmountString = ErgoAmount(element.text ?: 0).toStringRoundToDecimals(
-                    element.maxDecimals.toLong(),
+                    element.maxDecimals,
                     element.trimTrailingZero
                 )
